DC-14 Improve test function
* Move test file out of __tests_folder.

* Separate module for private function.

* Modify describe and test description.

* Create fixtures module for test read file.
