+++ packages/decoupled/jest.config.js
@@ -1,16 +1,14 @@
 module.exports = {
-  "roots": [
-    "<rootDir>/src"
-  ],
-  "transform": {
-    "^.+\\.ts?$": "ts-jest"
-  },
-  "testRegex": "(/__tests__/.*|(\\.|/)(test|spec))\\.ts?$",
-  "moduleFileExtensions": [
-    "ts",
-    "tsx",
-    "js",
-    "json",
-    "node"
-  ],
+    "roots": [
+        "<rootDir>/src"
+    ],
+    "transform": {
+        "^.+\\.ts?$": "ts-jest"
+    },
+    "testRegex": "(/__tests__/.*|(\\.|/)(test|spec))\\.ts?$",
+    "moduleFileExtensions": [
+        "ts",
+        "tsx",
+        "js"
+    ],
 }
