+++ packages/decoupled/package.json
@@ -55,6 +55,7 @@
   "devDependencies": {
     "jest": "^22.4.3",
     "regenerator-runtime": "^0.11.1",
-    "sinon": "^4.4.9"
+    "sinon": "^4.4.9",
+    "ts-jest": "^23.1.4"
   }
 }
