+++ packages/decoupled/src/fixtures/config/default/default.js
