+++ packages/decoupled/src/fixtures/config/example/example.js
