+++ packages/decoupled/src/fixtures/decoupled.json
@@ -1,0 +1,3 @@
+{
+  "appPath": "app"
+}
